.history/requirements_20241126213146.txt
# dependencies
beautifulsoup4
colorama
md2pdf
python-dotenv
pyyaml
uvicorn
pydantic
fastapi
python-multipart
markdown
langchain
langchain_community
langchain-openai
langchain-ollama
langgraph
tiktoken
gpt-researcher
arxiv
PyMuPDF
requests
jinja2
aiofiles
mistune
python-docx
htmldocx
lxml_html_clean
websockets
unstructured
json_repair
json5
loguru
groq>=0.4.0
pydantic>=2.0.0
langchain>=0.1.0
# uncomment for testing
# pytest
# pytest-asyncio
